you can select instruments and only hear what you select
